+++ download/download.go
@@ -8,32 +8,25 @@ import (
 	"time"
 )
 
+type getFilesConfig struct {
+	handler getURLsHandler
+	url     string
+}
+
 // Download all the files (might take several minutes).
 func Download(dir string, timeout time.Duration, urlsOnly, skip bool, parallel, retries int) error {
 	c := &http.Client{Timeout: timeout}
 	if !urlsOnly {
 		log.Output(2, "Preparing to download from the Federal Revenue official website…")
 	}
-
-	searches := []search{
-		{
-			"federal revenue",
-			federalRevenue,
-			federalRevenueSelector,
-			federalRevenueExtension,
-		},
-		{
-			"national treasure",
-			nationalTreasure,
-			nationalTreasureSelector,
-			nationalTreasureExtension,
-		},
+	confs := []getFilesConfig{
+		{federalRevenueGetURLs, federalRevenueURL},
+		{nationalTreasureGetURLs, nationalTreasureBaseURL},
 	}
-	fs, err := getFiles(c, searches, dir, skip)
+	fs, err := getFiles(c, confs, dir, skip)
 	if err != nil {
-		return err
+		return fmt.Errorf("error gathering resources for download: %w", err)
 	}
-
 	if urlsOnly {
 		urls := make([]string, 0, len(fs))
 		for _, f := range fs {
@@ -45,10 +38,9 @@ func Download(dir string, timeout time.Duration, urlsOnly, skip bool, parallel, 
 		}
 		return nil
 	}
-
 	d, err := newDownloader(c, fs, 2, 4)
 	if err != nil {
-		return err
+		return fmt.Errorf("error creating a downloader: %w", err)
 	}
 	return d.downloadAll()
 }
