+++ testdata/national-treasure.json
@@ -1,0 +1,1 @@
+{"help": "https://www.tesourotransparente.gov.br/ckan/api/3/action/help_show?name=package_show", "success": true, "result": {"license_title": "License not specified", "maintainer": "COSIS/NUAFI", "relationships_as_object": [], "private": false, "maintainer_email": "ti.stn@tesouro.gov.br", "num_tags": 2, "id": "abb968cb-3710-4f85-89cf-875c91b9c7f6", "metadata_created": "2019-07-18T11:17:48.973741", "metadata_modified": "2022-01-06T11:46:22.017789", "author": "COSIS", "author_email": "ti.stn@tesouro.gov.br", "state": "active", "version": "", "creator_user_id": "c49a1e4d-9b45-4900-854b-bc5ad9d276eb", "type": "dataset", "resources": [{"resource_group": "", "cache_last_updated": null, "package_id": "abb968cb-3710-4f85-89cf-875c91b9c7f6", "webstore_last_updated": null, "id": "eebb3bc6-9eea-4496-8bcf-304f33155282", "size": "435942", "state": "active", "hash": "", "description": "TABELA DE MUNIC\u00cdPIOS DO SIAFI", "format": "CSV", "tracking_summary": {"total": 106, "recent": 22}, "mimetype_inner": null, "url_type": "upload", "mimetype": null, "cache_url": null, "name": "TABELA DE MUNIC\u00cdPIOS DO SIAFI", "created": "2019-07-18T11:29:29.208001", "url": "https://www.tesourotransparente.gov.br/ckan/dataset/abb968cb-3710-4f85-89cf-875c91b9c7f6/resource/eebb3bc6-9eea-4496-8bcf-304f33155282/download/TABMUN.CSV", "webstore_url": null, "last_modified": "2021-12-21T15:17:17.535254", "position": 0, "revision_id": "23e77ca1-6c1d-453f-9a1d-b9e78a5b0894", "resource_type": null}], "num_resources": 1, "tags": [{"vocabulary_id": null, "state": "active", "display_name": "MUNIC\u00cdPIOS", "id": "7a33f2c5-faa7-46aa-a00e-859b737eebf6", "name": "MUNIC\u00cdPIOS"}, {"vocabulary_id": null, "state": "active", "display_name": "SIAFI", "id": "d8cf1bba-6399-49a1-940d-8f278045b222", "name": "SIAFI"}], "tracking_summary": {"total": 0, "recent": 0}, "groups": [], "license_id": "notspecified", "relationships_as_subject": [], "organization": {"description": "", "created": "2015-11-17T11:06:39.943667", "title": "COSIS", "name": "cosis", "is_organization": true, "state": "active", "image_url": "", "revision_id": "5c3b3359-ff7e-4f56-a98c-6f2d0d3700d6", "type": "organization", "id": "71e665c2-29e1-4f54-8e5b-278482a06ec9", "approval_status": "approved"}, "name": "lista-de-municipios-do-siafi", "isopen": false, "url": "", "notes": "Lista de munic\u00edpios do SIAFI com c\u00f3digo SIAFI e c\u00f3digo IBGE", "owner_org": "71e665c2-29e1-4f54-8e5b-278482a06ec9", "extras": [{"key": "Categoria do e-VoG", "value": "Economia e Finan\u00e7as"}, {"key": "Data de in\u00edcio dos dados", "value": ""}, {"key": "Frequ\u00eancia de Atualiza\u00e7\u00e3o", "value": "Anual"}, {"key": "resource_group", "value": ""}], "title": "Lista de munic\u00edpios do SIAFI", "revision_id": "d9fc132e-46f8-471e-90a0-87842250dbe4"}}
